+++ doc-delivery/src/main/java/com/wd/cloud/docdelivery/repository/LiteratureRepository.java
@@ -10,10 +10,22 @@ import java.util.List;
  * @date 2018/5/7
  * @Description:
  */
-public interface LiteratureRepository extends JpaRepository<Literature, Integer> {
+public interface LiteratureRepository extends JpaRepository<Literature, Long> {
 
+    /**
+     * 根据文献标题查询文献元数据
+     * @param docTitle
+     * @return
+     */
     Literature findByDocTitle(String docTitle);
 
+
+    /**
+     * 根据文献标题和文献连接查询元数据
+     * @param docTitle
+     * @param docHref
+     * @return
+     */
     Literature findByDocTitleAndDocHref(String docTitle,String docHref);
 
 }
